Combine test files.
